DRY and nomenclature corrections
