+++ lib/import/brca/core/genotype_brca.rb
@@ -10,7 +10,7 @@ module Import
       # negative ones)produced sequence variants.
       # However, each genotype also contains all the information available about test and result
       # level fields, so that the storage processor can match and create the appropriate tables
-      class GenotypeBrca < Import::Genotype
+      class GenotypeBrca < Import::Germline::Genotype
         BRCA_MAP = { 'BRCA1' => 7,
                      'BRCA2' => 8,
                      'ATM' => 451,
