+++ lib/import/brca/providers/leeds/report_extractor.rb
@@ -7,7 +7,7 @@ module Import
         module ReportExtractor
           # Parent class which encapsulates parsing logic associated with common Leeds
           # genotype forms
-          class GenotypePattern < Import::ExtractionUtilities::CommonUtility
+          class GenotypePattern < Import::Utility::ExtractionUtilities::CommonUtility
             def initialize
               @failed_report_parse_counter = 0
               @report_parse_attempt_counter = 0
