+++ lib/import/brca/providers/newcastle/newcastle_handler.rb
@@ -6,13 +6,12 @@ module Import
     module Providers
       module Newcastle
         # Process Newcastle-specific record details into generalized internal genotype format
-        class NewcastleHandler < Import::Brca::Core::ProviderHandler
+        class NewcastleHandler < Import::Germline::ProviderHandler
           include Import::Helpers::Brca::Providers::Rtd::RtdConstants
 
           def initialize(batch)
             @failed_variant_counter = 0
             @variants_processed_counter = 0
-            @ex = Import::ExtractionUtilities::LocationExtractor.new
             super
           end
