+++ lib/import/colorectal/providers/manchester/manchester_handler_colorectal.rb
@@ -4,7 +4,7 @@ module Import
     module Providers
       module Manchester
         # Manchester R0A importer
-        class ManchesterHandlerColorectal < Import::Brca::Core::ProviderHandler
+        class ManchesterHandlerColorectal < Import::Germline::ProviderHandler
           include Import::Helpers::Colorectal::Providers::R0a::R0aConstants
           include Import::Helpers::Colorectal::Providers::R0a::R0aHelper
           def initialize(batch)
