+++ test/lib/import/brca/core/raw_record_test.rb
@@ -20,6 +20,6 @@ class RawRecordTest < ActiveSupport::TestCase
       'rawtext_clinical.to_json' => {}.to_json
     }
 
-    Import::Brca::Core::RawRecord.new(default_options.merge!(options))
+    Import::Germline::RawRecord.new(default_options.merge!(options))
   end
 end
